Refactoring of the HTTP code
